Added jsapi (google Geochart), Simple world map and USA map with states added
I have included jsapi (geochart) from google to the project. Added the
lib to lib/jsapi/jsapi.html
Instructions to use this API:
“https://google-developers.appspot.com/chart/interactive/docs/gallery/ge
ochart”
Created a map.js that paints the USA map with states and the World Map
with countries
The maps are painted in two different containers USAmap and Worldmap.